# transaction-orchestrator/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Transaction Orchestrator API
  description: API for transaction orchestration
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1

tags:
  - name: Transactions

paths:

  /transactions:
    post:
      tags:
        - Transactions
      summary: Create a new transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponseWrapper'
        '400':
          description: Validation or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseWrapper'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseWrapper'
        '502':
          description: Provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseWrapper'

  /transactions/{transaction_id}:
    get:
      tags:
        - Transactions
      summary: Get transaction by ID
      operationId: getTransaction
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponseWrapper'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseWrapper'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseWrapper'
    patch:
      tags:
        - Transactions
      summary: Update transaction status
      operationId: updateTransaction
      parameters:
        - name: transaction_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionRequest'
      responses:
        '200':
          description: Transaction updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponseWrapper'
        '400':
          description: Invalid payload or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseWrapper'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseWrapper'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseWrapper'
        '502':
          description: Provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseWrapper'


components:

  schemas:

    # =========================
    # REQUESTS
    # =========================
    CreateTransactionRequest:
      type: object
      required:
        - client_transaction_id
        - amount
        - currency
        - country
        - payment_method_id
        - webhook_url
        - return_url
        - customer
      properties:
        client_transaction_id:
          type: string
          minLength: 1
          maxLength: 255
          example: "txn_123456"
        amount:
          type: integer
          minimum: 1
          description: Amount in cents
          example: 10000
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: "USD"
        country:
          type: string
          minLength: 2
          maxLength: 2
          example: "US"
        payment_method_id:
          type: string
          format: uuid
          example: "89132e2b-086d-4962-9b5a-7bb4607b2005"
        webhook_url:
          type: string
          pattern: "^https?://.+"
        return_url:
          type: string
          pattern: "^https?://.+"
        description:
          type: string
        expiration_time:
          type: integer
          description: Expiration time in seconds
        customer:
          $ref: '#/components/schemas/Customer'

    Customer:
      type: object
      required:
        - document_type
        - document_number
        - phone_code
        - phone_number
        - email
        - first_name
        - last_name
      properties:
        document_type:
          type: string
          minLength: 1
          maxLength: 20
          example: "CC"
        document_number:
          type: string
          minLength: 1
          maxLength: 50
          example: "123456789"
        phone_code:
          type: string
          pattern: '^\+?\d+$'
          example: "57"
        phone_number:
          type: string
          pattern: '^\d+$'
          example: "3001234567"
        email:
          type: string
          format: email
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        second_name:
          type: string
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        second_last_name:
          type: string

    UpdateTransactionRequest:
      type: object
      required:
        - transaction_id
        - status
      properties:
        transaction_id:
          type: string
          format: uuid
        provider_transaction_id:
          type: string
        status:
          $ref: '#/components/schemas/TransactionStatus'
        message:
          type: string
        processed_at:
          type: string
          format: date-time

    # =========================
    # RESPONSES
    # =========================
    TransactionResponseWrapper:
      type: object
      properties:
        code:
          type: string
          example: "000"
        message:
          type: string
          example: "Successful operation"
        data:
          $ref: '#/components/schemas/TransactionData'

    ErrorResponseWrapper:
      type: object
      properties:
        code:
          type: string
          example: "1001"
        message:
          type: string
          example: "Validation error"
        data:
          nullable: true

    GenericResponse:
      type: object
      properties:
        code:
          type: string
          example: "000"
        message:
          type: string
          example: "Webhook received"

    # =========================
    # DATA
    # =========================
    TransactionData:
      type: object
      properties:
        transaction_id:
          type: string
          format: uuid
        client_transaction_id:
          type: string
        payment_method_id:
          type: string
        amount:
          type: integer
        currency:
          type: string
        country:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TransactionStatus'
        created_at:
          type: string
          format: date-time

    # =========================
    # ENUMS
    # =========================
    TransactionStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - SUCCESS
        - FAILED